extract CachedFunc to utils
