+++ crates/polars-time/src/chunkedarray/utf8/mod.rs
@@ -7,41 +7,12 @@ use chrono::ParseError;
 pub use patterns::Pattern;
 #[cfg(feature = "dtype-time")]
 use polars_core::chunked_array::temporal::time_to_time64ns;
+use polars_utils::cache::CachedFunc;
 
 use super::*;
 #[cfg(feature = "dtype-date")]
 use crate::chunkedarray::date::naive_date_to_date;
 use crate::prelude::utf8::strptime::StrpTimeState;
-
-struct CachedFunc<T, R, F> {
-    func: F,
-    cache: PlHashMap<T, R>,
-}
-
-impl<T, R, F> CachedFunc<T, R, F>
-where
-    F: FnMut(T) -> R,
-    T: std::hash::Hash + Eq + Clone,
-    R: Copy,
-{
-    pub fn new(func: F) -> Self {
-        Self {
-            func,
-            cache: PlHashMap::new(),
-        }
-    }
-
-    pub fn eval(&mut self, x: T, use_cache: bool) -> R {
-        if use_cache {
-            *self
-                .cache
-                .entry(x)
-                .or_insert_with_key(|xr| (self.func)(xr.clone()))
-        } else {
-            (self.func)(x)
-        }
-    }
-}
 
 #[cfg(feature = "dtype-time")]
 fn time_pattern<F, K>(val: &str, convert: F) -> Option<&'static str>
